Address PR feedback for status command
