+++ src/lib.rs
@@ -1,2 +1,3 @@
 pub mod engine;
+pub mod status;
 pub mod ui;
